+++ cargo-insta/tests/functional/workspace.rs
@@ -571,7 +571,6 @@ fn test_hello() {
     ---
     source: "../tests/lib.rs"
     expression: hello()
-    snapshot_kind: text
     ---
     Hello, world!
     "#);