+++ insta/src/runtime.rs
@@ -23,14 +23,13 @@ use crate::{
     snapshot::TextSnapshotKind,
 };
 
-lazy_static::lazy_static! {
-    static ref TEST_NAME_COUNTERS: Mutex<BTreeMap<String, usize>> =
-        Mutex::new(BTreeMap::new());
-    static ref TEST_NAME_CLASH_DETECTION: Mutex<BTreeMap<String, bool>> =
-        Mutex::new(BTreeMap::new());
-    static ref INLINE_DUPLICATES: Mutex<BTreeSet<String>> =
-        Mutex::new(BTreeSet::new());
-}
+use once_cell::sync::Lazy;
+
+static TEST_NAME_COUNTERS: Lazy<Mutex<BTreeMap<String, usize>>> =
+    Lazy::new(|| Mutex::new(BTreeMap::new()));
+static TEST_NAME_CLASH_DETECTION: Lazy<Mutex<BTreeMap<String, bool>>> =
+    Lazy::new(|| Mutex::new(BTreeMap::new()));
+static INLINE_DUPLICATES: Lazy<Mutex<BTreeSet<String>>> = Lazy::new(|| Mutex::new(BTreeSet::new()));
 
 thread_local! {
     static RECORDED_DUPLICATES: RefCell<Vec<BTreeMap<String, Snapshot>>> = RefCell::default()