+++ insta/src/snapshot.rs
@@ -1,28 +1,27 @@
+use crate::{
+    content::{self, json, yaml, Content},
+    elog,
+    utils::style,
+};
+use once_cell::sync::Lazy;
+use std::env;
 use std::error::Error;
+use std::fmt;
 use std::fs;
 use std::io::{BufRead, BufReader, Write};
 use std::path::{Path, PathBuf};
 use std::rc::Rc;
 use std::time::{SystemTime, UNIX_EPOCH};
 use std::{borrow::Cow, iter::once};
-use std::{env, fmt};
 
-use crate::{
-    content::{self, json, yaml, Content},
-    elog,
-    utils::style,
-};
-
-lazy_static::lazy_static! {
-    static ref RUN_ID: String = {
-        if let Ok(run_id) = env::var("NEXTEST_RUN_ID") {
-            run_id
-        } else {
-            let d = SystemTime::now().duration_since(UNIX_EPOCH).unwrap();
-            format!("{}-{}", d.as_secs(), d.subsec_nanos())
-        }
-    };
-}
+static RUN_ID: Lazy<String> = Lazy::new(|| {
+    if let Ok(run_id) = env::var("NEXTEST_RUN_ID") {
+        run_id
+    } else {
+        let d = SystemTime::now().duration_since(UNIX_EPOCH).unwrap();
+        format!("{}-{}", d.as_secs(), d.subsec_nanos())
+    }
+});
 
 /// Holds a pending inline snapshot loaded from a json file or read from an assert
 /// macro (doesn't write to the rust file, which is done by `cargo-insta`)
@@ -291,15 +290,13 @@ impl MetaData {
             fields.push(("input_file", Content::from(input_file)));
         }
 
-        let snapshot_type = Content::from(match self.snapshot_kind {
-            SnapshotKind::Text => "text",
+        match self.snapshot_kind {
+            SnapshotKind::Text => {}
             SnapshotKind::Binary { ref extension } => {
                 fields.push(("extension", Content::from(extension.clone())));
-                "binary"
+                fields.push(("snapshot_kind", Content::from("binary")));
             }
-        });
-
-        fields.push(("snapshot_kind", snapshot_type));
+        }
 
         Content::Struct("MetaData", fields)
     }
@@ -385,7 +382,7 @@ impl Snapshot {
                     }
                 }
             }
-            elog!("A snapshot uses a legacy snapshot format; please update it to the new format with `cargo insta test --force-update-snapshots --accept`.\n\nSnapshot is at: {}", p.to_string_lossy());
+            elog!("A snapshot uses a legacy snapshot format; please update it to the new format with `cargo insta test --force-update-snapshots --accept`.\nSnapshot is at: {}", p.to_string_lossy());
             rv
         };
 
@@ -566,6 +563,9 @@ impl Snapshot {
         buf
     }
 
+    // We take `md` as an argument here because the calling methods want to
+    // adjust it; e.g. removing volatile fields when writing to the final
+    // `.snap` file.
     fn save_with_metadata(&self, path: &Path, md: &MetaData) -> Result<(), Box<dyn Error>> {
         if let Some(folder) = path.parent() {
             fs::create_dir_all(folder)?;
@@ -592,9 +592,6 @@ impl Snapshot {
     }
 
     /// Saves the snapshot.
-    ///
-    /// Returns `true` if the snapshot was saved.  This will return `false` if there
-    /// was already a snapshot with matching contents.
     #[doc(hidden)]
     pub fn save(&self, path: &Path) -> Result<(), Box<dyn Error>> {
         self.save_with_metadata(path, &self.metadata.trim_for_persistence())